gitignore added
