+++ alymov/lab_8/README.md
@@ -1,0 +1,14 @@
+
+## Lab 8
+
+
+usage
+```bash
+   ./main.py
+```
+help
+```bash
+    help: ./main.py --help
+```
+
+![Alt text](sol.png)
