+++ alymov/lab_8/lab_8.cpp
@@ -128,3 +128,24 @@ void embed_vector(Matrix& A, Vector& b, bool cols, int ind)
     }
     return;
 }
+
+void save_to_file(std::vector<Matrix>& system)
+{
+    std::ofstream file("sol.csv");
+
+    for(int cnt=0; cnt<system.size(); cnt++){
+        Matrix& sol = system[cnt];
+        for(int k=0; k<sol.rows; k++)
+        {
+            for(int j=0; j<sol.cols; j++)
+            {
+                file << sol[k][j];
+                if(j!=sol.cols-1)
+                    file << ",";
+            }
+            file << "\n";
+        }
+    }
+    file.close();
+    return;
+}
