+++ alymov/lab_8/sol.csv
@@ -4,69 +4,27 @@
 -0.309017,-0.25,-0.0954915,0.0954916,0.25,0.309017
 -0.809017,-0.654509,-0.25,0.25,0.654509,0.809017
 -1,-0.809017,-0.309017,0.309017,0.809017,1
-0.983471,0.795645,0.303909,-0.30391,-0.795645,-0.983471
-0.795645,0.644066,0.246013,-0.246013,-0.644066,-0.795645
-0.303909,0.246015,0.0939699,-0.09397,-0.246015,-0.303909
--0.30391,-0.246015,-0.09397,0.0939701,0.246015,0.30391
--0.795645,-0.644066,-0.246013,0.246013,0.644066,0.795645
--0.983471,-0.795645,-0.303909,0.30391,0.795645,0.983471
-0.967216,0.782494,0.298886,-0.298886,-0.782494,-0.967216
-0.782494,0.633772,0.242086,-0.242086,-0.633772,-0.782494
-0.298886,0.242093,0.0924739,-0.092474,-0.242093,-0.298886
--0.298886,-0.241886,-0.0923946,0.0923947,0.241886,0.298886
--0.782494,-0.623717,-0.238245,0.238245,0.623717,0.782494
--0.967216,-0.782494,-0.298886,0.298886,0.782494,0.967216
-0.951229,0.769561,0.293946,-0.293946,-0.769561,-0.951229
-0.769561,0.623626,0.238218,-0.238218,-0.623626,-0.769561
-0.293946,0.238238,0.0910043,-0.0910044,-0.238238,-0.293946
--0.293946,-0.23765,-0.0907762,0.0907763,0.23765,0.293946
--0.769561,-0.6047,-0.230903,0.230903,0.6047,0.769561
--0.951229,-0.769561,-0.293946,0.293946,0.769561,0.951229
-0.935507,0.756841,0.289087,-0.289088,-0.756841,-0.935507
-0.756841,0.613627,0.234408,-0.234408,-0.613627,-0.756841
-0.289087,0.234449,0.0895616,-0.0895617,-0.234449,-0.289087
--0.289088,-0.23334,-0.0891244,0.0891245,0.23334,0.289088
--0.756841,-0.586892,-0.223959,0.223959,0.586892,0.756841
--0.935507,-0.756841,-0.289087,0.289088,0.756841,0.935507
-0.920044,0.744332,0.284309,-0.28431,-0.744332,-0.920044
-0.744332,0.603774,0.230659,-0.230466,-0.594434,-0.744332
-0.284309,0.230729,0.0881475,-0.0880721,-0.227073,-0.284309
--0.28431,-0.228983,-0.0874498,0.0873749,0.225355,0.28431
--0.744332,-0.570185,-0.217391,0.217209,0.561375,0.744332
--0.920044,-0.744332,-0.284309,0.28431,0.744332,0.920044
-0.904837,0.732029,0.27961,-0.27961,-0.732029,-0.904837
-0.732029,0.594067,0.226974,-0.226422,-0.576189,-0.732029
-0.27961,0.227077,0.0867635,-0.0865441,-0.220001,-0.27961
--0.27961,-0.224604,-0.0857615,0.0855445,0.217606,0.27961
--0.732029,-0.55448,-0.211179,0.210665,0.537837,0.732029
--0.904837,-0.732029,-0.27961,0.27961,0.732029,0.904837
-0.889882,0.719929,0.274988,-0.274989,-0.71993,-0.889882
-0.719929,0.584505,0.223351,-0.222484,-0.567893,-0.719929
-0.274988,0.223492,0.0854091,-0.0850576,-0.216768,-0.274988
--0.274989,-0.220223,-0.0840664,0.0837202,0.213601,0.274989
--0.71993,-0.539689,-0.205296,0.204501,0.524443,0.71993
--0.889882,-0.719929,-0.274988,0.274989,0.71993,0.889882
-0.875173,0.70803,0.270443,-0.270444,-0.70803,-0.875173
-0.70803,0.575087,0.219787,-0.218646,-0.559641,-0.70803
-0.270443,0.219975,0.0840836,-0.0836114,-0.213589,-0.270443
--0.270444,-0.215856,-0.0823705,0.0819077,0.209597,0.270444
--0.70803,-0.525732,-0.19972,0.198688,0.511752,0.70803
--0.875173,-0.70803,-0.270443,0.270444,0.70803,0.875173
-0.860708,0.696327,0.265973,-0.265974,-0.696328,-0.860708
-0.696327,0.565811,0.216284,-0.214738,-0.543325,-0.696327
-0.265973,0.216523,0.0827877,-0.0821402,-0.207287,-0.265973
--0.265974,-0.211519,-0.0806805,0.080049,0.202506,0.265974
--0.696328,-0.512538,-0.194434,0.193052,0.492369,0.696328
--0.860708,-0.696327,-0.265973,0.265974,0.696328,0.860708
-0.846482,0.684818,0.261577,-0.261577,-0.684818,-0.846482
-0.684818,0.556677,0.212843,-0.21078,-0.527748,-0.684818
-0.261577,0.213137,0.0815217,-0.0806499,-0.201233,-0.261577
--0.261577,-0.207222,-0.0790022,0.0781563,0.195666,0.261577
--0.684818,-0.500041,-0.189418,0.187595,0.474345,0.684818
--0.846482,-0.684818,-0.261577,0.261577,0.684818,0.846482
-0.832491,0.673499,0.257254,-0.257254,-0.673499,-0.832491
-0.673499,0.547684,0.209466,-0.20679,-0.512856,-0.673499
-0.257254,0.209814,0.0802857,-0.0791457,-0.195416,-0.257254
--0.257254,-0.202977,-0.0773404,0.0762407,0.189075,0.257254
--0.673499,-0.488184,-0.184657,0.182317,0.457548,0.673499
--0.832491,-0.673499,-0.257254,0.257254,0.673499,0.832491
+0.960789,0.777295,0.2969,-0.2969,-0.777295,-0.960789
+0.777295,0.629807,0.240574,-0.240574,-0.629807,-0.777295
+0.2969,0.240584,0.0918989,-0.0918989,-0.240584,-0.2969
+-0.2969,-0.240585,-0.0918989,0.091899,0.240585,0.2969
+-0.777295,-0.629807,-0.240574,0.240575,0.629807,0.777295
+-0.960789,-0.777295,-0.2969,0.2969,0.777295,0.960789
+0.923116,0.746817,0.285259,-0.285259,-0.746817,-0.923116
+0.746817,0.605938,0.231481,-0.231481,-0.605938,-0.746817
+0.285259,0.231543,0.0884554,-0.0884554,-0.231543,-0.285259
+-0.285259,-0.230532,-0.0880693,0.0880694,0.230532,0.285259
+-0.746817,-0.584976,-0.223473,0.223473,0.584976,0.746817
+-0.923116,-0.746817,-0.285259,0.285259,0.746817,0.923116
+0.88692,0.717534,0.274073,-0.274074,-0.717534,-0.88692
+0.717534,0.582893,0.222711,-0.222712,-0.582893,-0.717534
+0.274073,0.222884,0.0851646,-0.0851646,-0.222884,-0.274073
+-0.274074,-0.220234,-0.0841172,0.0841173,0.220234,0.274074
+-0.717534,-0.546392,-0.208417,0.208417,0.546392,0.717534
+-0.88692,-0.717534,-0.274073,0.274074,0.717534,0.88692
+0.852144,0.689399,0.263327,-0.263327,-0.689399,-0.852144
+0.689399,0.56066,0.214256,-0.214256,-0.56066,-0.689399
+0.263327,0.214602,0.082025,-0.0820251,-0.214602,-0.263327
+-0.263327,-0.209962,-0.080134,0.0801341,0.209962,0.263327
+-0.689399,-0.512845,-0.195113,0.195113,0.512845,0.689399
+-0.852144,-0.689399,-0.263327,0.263327,0.689399,0.852144
